[lib] Use ‘size_t’ for ‘svz_hash_t’ dimensions.
U ≡ "Convert type to ‘size_t’ for local var";
S(X) ≡ "Likewise, for struct member(s) X".
* src/libserveez/hash.c (svz_hash_key_length): Return ‘size_t’; U.
(svz_hash_create, svz_hash_destroy): U.
(svz_hash_rehash): U; remove cast to ‘unsigned long’.
(svz_hash_foreach): U.
(svz_hash_size): Return ‘size_t’.
(svz_hash_contains): U.
* src/libserveez/hash.h (svz_hash_t): S (buckets, fill, keys, keylen).
(svz_hash_create, svz_hash_size): Update func decls.

U ≡ "Convert type to ‘size_t’ for local var";
S(X) ≡ "Likewise, for struct member(s) X".
* src/http-server/http-cache.c
(http_cache_entries): Convert type to ‘size_t’.
(http_alloc_cache): U.
* src/http-server/http-cache.h
(http_cache_entries): Update var decl.
(http_alloc_cache): Update func decl.
* src/http-server/http-proto.c (http_init): U.
* src/irc-server/irc-proto.c (irc_regex_channel, irc_regex_nick): U.
* src/nut-server/gnutella.h (nut_config_t): S (connections).

+++ src/http-server/http-cache.c
@@ -50,7 +50,7 @@
 #include "http-cache.h"
 
 svz_hash_t *http_cache = NULL;               /* actual cache entry hash */
-int http_cache_entries = 0;                  /* amount of cache entries */
+size_t http_cache_entries = 0;               /* amount of cache entries */
 http_cache_entry_t *http_cache_first = NULL; /* most recent entry */
 http_cache_entry_t *http_cache_last = NULL;  /* least recent entry */
 
@@ -58,7 +58,7 @@ http_cache_entry_t *http_cache_last = NULL;  /* least recent entry */
  * This will initialize the http cache entries.
  */
 void
-http_alloc_cache (int entries)
+http_alloc_cache (size_t entries)
 {
   if (entries > http_cache_entries || http_cache == NULL)
     {
